Refactored form-footer and added basic footer.
Moved form-footer into base-elements and created a basic footer.
Refactored the footer color variables to juat be __footer instead of
__form-footer.  I added a readme for the basic footer but I'm not sure
about the content to write.  Also, the basic footer is underneath the
form-footer since they are both fixed to the bottom.  Form-footer is
responsive-ish.  Some of the elements move out of the footer onto the
page when resized (image attached to show).
Resized checkbox element by Tessa's request.