modal should be visibility: hidden when hidden
It’s possible to tab into the elements (close buttons, other buttons,
whatever) of a modal when it’s not visible. This fixes that.

nudged the `.modal-visible` class to be closer to the `.modal` class
that it modifies. Should it be `.modal--is-visible`?

I also noticed that you’re not actually focusing on the modal when it’s
opened. you were checking height/width of modalElement, which is an
array of elements. What you really want is either modalElement[0] or
modal.element