Update typescale to be mathmatically correct (#1156)
Some of the typescale rem values were incorrect (i.e. 12/16 is .75 not .79). Other values were being rounded to 3 decimal places. We've extended this to 4 decimal places in for a few values because...

The difference in browser support is negligible.
These values are now mathematically correct and future proof against better browser rounding.