fix(global scss vars): check for global variables and booleans to prevent import errors (#1912)
* fix(reset): declare $use-reset var in reset.scss

This fixes a situation when User imports reset.scss by itself, they would encounter an error
because of the missing variable, which was only in styles.scss.

* fix(body): set $use-globals variable in body.scss

* fix(font-face): set $use-fonts variable

* refactor(sass vars): use global-variable-exists() function

Sass function that checks if a global variable exists with the same
name. Using this allows us to avoid multiple declarations of the same
variable.

* refactor(typography.scss): wrap all typography styles in mixin

Matches new (but non-breaking) changes proposed in [7.x] Refactor global
scss #1891 PR. This makes it configurable for typography CSS to be
included or not. This is included by default.