feat(build): use ES dynamic import syntax for conditional imports (#416)
The new code uses ES dynamic import so that the import works with run-time conditionals. Run-time conditional is needed for an environment where no static build is in use (e.g. server-side rendering).

Such "dynamic" nature requires us to do additional processing before the code is passed along to WebPack to make sure Sass import does not run when `EXCLUDE_SASS` is set - Which is, removing the Sass import code along with the `if ()` clause. README is updated accordingly.

Fixes #415.