fix(checkbox): changed prop type for labeltext
e.g. the checkbox labelText is limited to be used with the string proptypes, so there will be warnings when passing in a number, or a valid component. 

As no modifications happen on the props and this is only something that gets rendered down the line, why limit its use? Use the proptype.node instead of proptype.string, proptype.number to allow anything that can be rendered to be passed in.

Everywhere in the code where props are simply being rendered without being modified, change the proptype from e.g. proptype.string, proptype.number to proptype.node