upgrade express to address vulnerabilities (#698)
snyk test express@4.13.4

Testing express@4.13.4...
✗ Low severity vulnerability found on debug@2.2.0
- desc: Regular Expression Denial of Service (ReDoS)
- info: https://snyk.io/vuln/npm:debug:20170905
- from: express@4.13.4 > debug@2.2.0
You've tested an outdated version of the project. Should be upgraded to express@4.15.5 (triggers upgrades to debug@2.6.9)

✗ Low severity vulnerability found on debug@2.2.0
- desc: Regular Expression Denial of Service (ReDoS)
- info: https://snyk.io/vuln/npm:debug:20170905
- from: express@4.13.4 > finalhandler@0.4.1 > debug@2.2.0
You've tested an outdated version of the project. Should be upgraded to express@4.15.0 (triggers upgrades to finalhandler@1.0.6 > debug@2.6.9)

✗ Low severity vulnerability found on debug@2.2.0
- desc: Regular Expression Denial of Service (ReDoS)
- info: https://snyk.io/vuln/npm:debug:20170905
- from: express@4.13.4 > send@0.13.1 > debug@2.2.0
You've tested an outdated version of the project. Should be upgraded to express@4.15.5 (triggers upgrades to send@0.15.6 > debug@2.6.9)

✗ Low severity vulnerability found on debug@2.2.0
- desc: Regular Expression Denial of Service (ReDoS)
- info: https://snyk.io/vuln/npm:debug:20170905
- from: express@4.13.4 > serve-static@1.10.3 > send@0.13.2 > debug@2.2.0
You've tested an outdated version of the project. Should be upgraded to express@4.15.5 (triggers upgrades to serve-static@1.12.6 > send@0.15.6 > debug@2.6.9)

✗ High severity vulnerability found on fresh@0.3.0
- desc: Regular Expression Denial of Service (ReDoS)
- info: https://snyk.io/vuln/npm:fresh:20170908
- from: express@4.13.4 > fresh@0.3.0
You've tested an outdated version of the project. Should be upgraded to express@4.15.5 (triggers upgrades to fresh@0.5.2)

✗ High severity vulnerability found on fresh@0.3.0
- desc: Regular Expression Denial of Service (ReDoS)
- info: https://snyk.io/vuln/npm:fresh:20170908
- from: express@4.13.4 > send@0.13.1 > fresh@0.3.0
You've tested an outdated version of the project. Should be upgraded to express@4.15.5 (triggers upgrades to send@0.15.6 > fresh@0.5.2)

✗ High severity vulnerability found on fresh@0.3.0
- desc: Regular Expression Denial of Service (ReDoS)
- info: https://snyk.io/vuln/npm:fresh:20170908
- from: express@4.13.4 > serve-static@1.10.3 > send@0.13.2 > fresh@0.3.0
You've tested an outdated version of the project. Should be upgraded to express@4.15.5 (triggers upgrades to serve-static@1.12.6 > send@0.15.6 > fresh@0.5.2)

✗ Low severity vulnerability found on mime@1.3.4
- desc: Regular Expression Denial of Service (ReDoS)
- info: https://snyk.io/vuln/npm:mime:20170907
- from: express@4.13.4 > send@0.13.1 > mime@1.3.4
You've tested an outdated version of the project. Should be upgraded to express@4.16.0 (triggers upgrades to send@0.16.0 > mime@1.4.1)

✗ Low severity vulnerability found on mime@1.3.4
- desc: Regular Expression Denial of Service (ReDoS)
- info: https://snyk.io/vuln/npm:mime:20170907
- from: express@4.13.4 > serve-static@1.10.3 > send@0.13.2 > mime@1.3.4
You've tested an outdated version of the project. Should be upgraded to express@4.16.0 (triggers upgrades to serve-static@1.13.0 > send@0.16.0 > mime@1.4.1)

✗ Low severity vulnerability found on ms@0.7.1
- desc: Regular Expression Denial of Service (ReDoS)
- info: https://snyk.io/vuln/npm:ms:20170412
- from: express@4.13.4 > debug@2.2.0 > ms@0.7.1
You've tested an outdated version of the project. Should be upgraded to express@4.15.3 (triggers upgrades to debug@2.6.7 > ms@2.0.0)

✗ Low severity vulnerability found on ms@0.7.1
- desc: Regular Expression Denial of Service (ReDoS)
- info: https://snyk.io/vuln/npm:ms:20170412
- from: express@4.13.4 > finalhandler@0.4.1 > debug@2.2.0 > ms@0.7.1
You've tested an outdated version of the project. Should be upgraded to express@4.15.0 (triggers upgrades to finalhandler@1.0.3 > debug@2.6.7 > ms@2.0.0)

✗ Low severity vulnerability found on ms@0.7.1
- desc: Regular Expression Denial of Service (ReDoS)
- info: https://snyk.io/vuln/npm:ms:20170412
- from: express@4.13.4 > send@0.13.1 > debug@2.2.0 > ms@0.7.1
You've tested an outdated version of the project. Should be upgraded to express@4.15.3 (triggers upgrades to send@0.15.3 > debug@2.6.7 > ms@2.0.0)

✗ Low severity vulnerability found on ms@0.7.1
- desc: Regular Expression Denial of Service (ReDoS)
- info: https://snyk.io/vuln/npm:ms:20170412
- from: express@4.13.4 > serve-static@1.10.3 > send@0.13.2 > debug@2.2.0 > ms@0.7.1
You've tested an outdated version of the project. Should be upgraded to express@4.15.3 (triggers upgrades to serve-static@1.12.3 > send@0.15.3 > debug@2.6.7 > ms@2.0.0)

✗ Low severity vulnerability found on ms@0.7.1
- desc: Regular Expression Denial of Service (ReDoS)
- info: https://snyk.io/vuln/npm:ms:20170412
- from: express@4.13.4 > send@0.13.1 > ms@0.7.1
You've tested an outdated version of the project. Should be upgraded to express@4.15.3 (triggers upgrades to send@0.15.3 > ms@2.0.0)

✗ Low severity vulnerability found on ms@0.7.1
- desc: Regular Expression Denial of Service (ReDoS)
- info: https://snyk.io/vuln/npm:ms:20170412
- from: express@4.13.4 > serve-static@1.10.3 > send@0.13.2 > ms@0.7.1
You've tested an outdated version of the project. Should be upgraded to express@4.15.3 (triggers upgrades to serve-static@1.12.3 > send@0.15.3 > ms@2.0.0)

✗ High severity vulnerability found on negotiator@0.5.3
- desc: Regular Expression Denial of Service (DoS)
- info: https://snyk.io/vuln/npm:negotiator:20160616
- from: express@4.13.4 > accepts@1.2.13 > negotiator@0.5.3
You've tested an outdated version of the project. Should be upgraded to express@4.14.0 (triggers upgrades to accepts@1.3.3 > negotiator@0.6.1)

✗ High severity vulnerability found on qs@4.0.0
- desc: Prototype Override Protection Bypass
- info: https://snyk.io/vuln/npm:qs:20170213
- from: express@4.13.4 > qs@4.0.0
You've tested an outdated version of the project. Should be upgraded to express@4.15.2 (triggers upgrades to qs@6.4.0)