chore(components): rename --experimental mix-ins
When we release next version of Carbon, code in `componentname-experimental` will be the latest, stable code.
Which means `componentname-experimental` no longer makes sense given it's no longer "experimental".
This change is for renaming those to `componentname-x` so we don't have to go through big rename effort
right before releasing next version of Carbon.