chore(build): make dev Sass build faster
We used Sass importer because our dev env had `carbon-addons-cloud`
code with circular dependency back to this repository
(`carbon-components`). However, we found that it caused significant
slowdown of our Sass dev build to an unacceptable level for
change-style-and-see workflow.

This change puts a copy of such `carbon-addons-cloud` code in this
repository (as a dev-env-specific code) so we no longer need Sass
importer, which yields 5x `sass:dev` task performance.

In future, we should migrate from that code to UI shell side nav.

Refs #1502 and #1709.