fix(data-table): avoid clipping toolbar content
This change avoids clipping table toolbar unless batch action is
enabled, so that we allow non-floating menus to be placed in there.

Note that clipping table toolbar is still needed for the animation of
batch actions, though. Given that, we recommend not putting
non-floating menus in table toolbar if batch action is enbaled.

Fixes #3271.