fix(TableToolbarSearch): support back-tab
This change fixes the issue with `<TableToolbarSearch>` that happens
when the `<input>` in it has focus and user performs back-tab gesture
(Shift-Tab key).

In such scenario (in before-change version), the root element of
`<TableToolbarSearch>`
(`<div class="bx--toolbar-search-container-expandable">`) gets focus
given it has `tabindex="0"`, and the `onFocus()` handler of the element
attempts to send focus back to the `<input>` no matter what.

This change also fixes the following found in the debugging effort:

* An issue where `id` of `<Search>` in `<TableToolbarSearch>` is
  regenerated for every render
* Issues focus on `<input>` happening:
  * For every render, regardless of change in expanded state
  * Even for controlled change in expanded state (without user gesture)

Fixes #3762.