fix(OverflowMenu): support automatic primary focus
This change ports the logic introduced in #3476 that auto-detects an
overflow menu item to put focus on, without needing to specify
`data-floating-menu-primary-focus` in the menu item.

This change introduces `selectorFocusable` setting, for finding
focusable DOM elements. The formerly-used `selectorTabbable` is for
finding sequential-focus targets, which was not necessary the the right
list for finding the menu item to focus on.

This change also fixes an issue in `OverflowMenu-story.js` where the
menu could not be opened by mouse, etc.

Refs #3372.