chore(data-table): mark sticky header experimental (#3949)
In #3876, @jnm2377 showed some concerns wrt the complete-ness of sticky
header styles. Though we are aware that some teams have evaluated or
even started using sticky header styles, we see that it makes sense to
mark sticky header style as experimental in order to bring the concern
of complete-ness to users' attention.