fix(Slider): use value from input as-is
This change ensures that `<Slider>` won't force changing the value that
user types in the `<input>`. For example, if we set 3 to `min` and user
wants to type 100 in the `<input>`, this change ensures `<input>` won't
become 3 (the minimum value) as soon as user types 1 (the first
character in 100).

Fixes #3107.