fix(FloatingMenu): lazy evaluate trigger button position
This change switches the position calculation logic for the trigger
buttons of `<OverflowMenu>` and `<Tooltip>` from eager-evaluation on
mount to lazy-evaluation on rendering `<FloatingMenu>`.

This change relegates such logic from `<OverflowMenu>` and `<Tooltip>`
to `<FloatingMenu>` to do that.

Such change has two merits:

1. Avoids stale trigger position data, which happened when `<Tooltip>`
   is opened programmatically
2. Reduces `getBoundingClientRect()` calls when application is
   initialized

Also, swiches `<FloatingMenu>`'s prop for the DOM element ref of
trigger button, as well as the corresponding internal properties of
`<OverflowMenu>`/`<Tooltip>` from raw element to React ref.

Fixes #5305.