feat(icons-react): swap next for es, lib
This commit updates our icons-react package in order to remove our old bundling
strategy and swap it with the direction we're taking for v11.

BREAKING CHANGE: the icon component format has changed, we no longer export
sized icons and instead each icon supports a `size` prop. The exports from this
package have now changed to only be the icon name