feat(icons-react): swap next for es, lib (#11047)
* feat(icons-react): swap next for es, lib

This commit updates our icons-react package in order to remove our old bundling
strategy and swap it with the direction we're taking for v11.

BREAKING CHANGE: the icon component format has changed, we no longer export
sized icons and instead each icon supports a `size` prop. The exports from this
package have now changed to only be the icon name

* test(icons-react): update public api snapshot

* test(icons-react): update e2e tests for icons to use moduleInfo

* feat(react): update icons-react re-export entrypoint

* feat(upgrade): update transform to use icons-react path

* refactor(react): run icons codemod to update to new format

* test(e2e): update icons snapshot

* docs(react): update button story to remove 16px asset

* fix(react): update ChevronDown16 to ChevronDown

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>