fix(react/pagination): render page number twice when `pagesUnknown` (#12302)
* fix(react/pagination): change style of pageText when pagesUnknown is

According to #12252, when pagesUnknown is , pagination render page
number as same as selector which does not make sense to render this information twice.
So, I change page number (pageText) to page label in front of
the selector which makes more sense and make pagination looks cleaner.

BREAKING CHANGE: typeof pageText change from function to string.

* fix(react/pagination): render page number twice modified

* Update packages/react/src/components/Pagination/next/Pagination.js

Co-authored-by: Taylor Jones <tay1orjones@users.noreply.github.com>

* fix(react/pagination): resolve another conflict on /Pagination.js

* chore: fix deleted files

Co-authored-by: Taylor Jones <tay1orjones@users.noreply.github.com>
Co-authored-by: Taylor Jones <taylor.jones826@gmail.com>
Co-authored-by: TJ Egan <tw15egan@gmail.com>
Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>