fix(typescript): fix InlineCheckbox property types
Because InlineCheckbox does not have type definitions, it gets its input
property type definitions from React.forwardRef, and so does not allow extra
properties. This results in errors like:
Property 'ariaLabel' does not exist on type 'IntrinsicAttributes &
RefAttributes<any>'.
when we adopt TS in files that use InlineCheckbox. Since InlineCheckbox is
internal-only, policy is not to add types but to mark as type: any.