fix(typescript): fix DataTable exports tests
DataTable has a set of tests which check that all files in the directory (with
some specific exceptions) export a default component with the same name as the
file. This test got the expected name of the component by stripping ".js" from
the filename. This won't work for Typescript files, which have a different
extension. Stripped off the extension regardless of what it was instead.