Add Typescript annotations to TableSelectRow (#13042)
* refactor(NumberInput): add typescript annotations

Convert existing propType definitions to Typescript annotations on the
NumberInput component. This is part of a broader effort to add
Typescript annotations to components, tracked in #12513. Closes #12548.

Type annotation changes only; no breaking feature changes.

* fix(typescript): disable no-non-null when safe

Disable eslint no-no-null-assertion rule on a line where it is known to be safe
(because it will ultimately pass in a key-value pair of undefined: undefined,
which will have no effect on the class names produced).

* fix(typescript): fix docs and reinstante no-non-null

Fix doc for checked, which was incorrect in the propTypes.
Instead of disabling no-non-null and using an assertion,
check whether className is defined before adding it.

---------

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>