fix(lightbox-carousel): accessibility & QA improvements to lightbox carousels (#9149)
## Related Ticket(s)

Closes: #8911 #8913 #8915

Also related: #8912 #8914


## Description

This PR makes several accessibility improvements to the DDSCarousel and DDSLightboxMediaViewer components, as well as the lightbox carousel story

### #8911 Update Lightbox to use semantic headings ✅
Lightbox titles now render title text within `<h2 style="all: inherit;">` tags. This gives them semantic `h2` headers but does not bring along the user agent/carbon styles for the element.

### #8912 Add label for modal dialog ✅
While elements with `[role="dialog"]` are required to have a label for WCAG, that label should be relevant to the content within the dialog. Because the component cannot know what content (or even what language) is in it, all we can do is make space for content entry. This already exists within the component in the form of an `aria-labelledby` attribute pointing to [a div that accepts slotted content](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/blob/main/packages/web-components/src/components/expressive-modal/expressive-modal.ts#L358). This PR adds a visually hidden modal title to the lightbox carousel story

### #8913 Announce carousel navigation to screen reader ✅ 
The carousel's `renderStatus()` method now renders a visually-hidden span with `aria-live="polite"` and uses a new `formatAnnouncement` property function to format that announcement into a string that accounts for single slides or slide groups. This property is written adjacent to the `formatStatus` property that allows localizing the pagination information.

### #8914 Create new landmark region for carousels ✅ 
_~No changes were made for this ticket.~_

~8914 suggests, "to provide more context to the screen reader user, the carousel which includes the slides and the Carousel controls could be given in a labelled landmark region." While this would be good, wouldn't adding a landmark region require a label that describes _what is in_ the carousel, not just that it _is_ a carousel? It would not be difficult to add a fallback label if this is not the case.~

The carousel now has an additional `<div>` wrapping the scrolling contents and the navigation. This div has `role="region"` and an `aria-labelledby` attribute that points to a new div with a `<slot name="title"><span class="bx--visually-hidden">Carousel</span></slot>`. This allows authors to provide a title relevant to the carousel's contents with the generic fallback label of "carousel"

### #8915 Update carousel's next/previous button labels ✅
The carousel allows authors to specify the labels of the previous & next buttons, but we also provide a default. Previously this value was "next page", but the word `page` can be confusing. We now provide a default fallback that looks at the carousel's `pageSize` and returns either "next slide" or "next slide group" if more than one slide is visible.

### [No Ticket] Refactor carousel item's interactivity based on inert
When the `[inert]` attribute was added to the carousel, it was added in a way that wasn't implemented fully in safari/firefox, and didn't account for carousels with multiple visible values well. That's been updated to now use an intersection observer, and it will mark carousel items more less than 75% visible as `inert` and `aria-hidden`.


### Changelog

**Changed**

- Accessibility improvements to lightbox media viewer component
- Accessibility improvements to lightbox carousel story
- Accessibility improvements to carousel component