fix(ClickableTile): duplicate onKeyDown calls
Shouldnt call onKeyDown callback twice when enter or space is pressed.

If a user defines their own callback for onKeyDown, then there is a good chance they want to control what happens when any key is pressed (including enter and space). So preventDefault should be called for that use case.