fix: TableToolbarSearch parameter types (#15422)
* fix: tabletoolbarsearch parameters

Fix TableToolbarSearch parameters, in particular missing closeButtonLabelText.

Also refactor the code as done in @types/carbon-components-react:
Make TableToolbarSearch parameters extend SearchParameters.

IMO the TableToolbarSearch parameters have a lot of unnecessary differences
from the Search parameters, but I didn't change that in this commit.

One open question is why the initial expandedState is based on defaultValue.
That seems like a mistake.  If it is, I can roll it back in this PR.

Refs #14471.

* chore: remove todo, i guess current code makes sense

---------

Co-authored-by: TJ Egan <tw15egan@gmail.com>