Move watsonx app icons to AI category
Moving watsonx app icons to the AI category from the App Catalogue category. Per the request of @mjabbink

Note: Seems like if an icon is listed in multiple categories it will only show up in one on IDL/Carbon icon library. I chose to prioritize the AI category for now