fix(DataTable): Use the right heading level in TableContainer. (#18871)
* fix(TableContainer): use the right heading level

Assuming the application is already correctly using <Section>, this will
give the table the right heading level (i.e. one level higher than the
parent heading).

We could also add aria-labelledby={titleId} to the <Section> to give it an
implicit role of "region" rather than "presentation".  Not sure if a single
table is big enough to be considered a region or not.

Fixes #18835.

* fix(TableContainer): fix snapshot

* fix(TableContainer): second try to fix snapshot