fix(MenuButton): do not steal focus on keyboard click of MenuItem (#19120)
* fix(MenuButton): do not steal focus on keyboard click of MenuItem

Closes #18418.

When you click a MenuItem, the Menu closes and focus goes back to the MenuButton.
This makes sense except when the MenuItem's onClick() callback already focused
somewhere else.  The typical case is when the MenuItem opened a Modal or Tearsheet.

Actually, this was already working when the user clicked the MenuItem with the mouse.
The MenuButton was only stealing focus when the MenuItem was clicked via keyboard.
In that case, MenuItem's onClick() ran on keydown, and the Menu closed on keydown,
but the refocus() call didn't happen until keyup.

This PR changes MenuItem so the "click" happens on "keyup" rather than on "keydown".
So nothing happens on keydown, but on keyup, the following events happen in this order:

1. Refocus the MenuButton.
2. Close the Menu dropdown.
3. Execute onClick() callback, which may (or may not) focus somewhere else.

Rationale: When working on dijit, I found that "click" event on native elements like
BUTTON happened on "keyup", so it made sense to do it that way for DIV too, see
https://github.com/dojo/dijit/blob/master/a11yclick.js.

I tested by modifying MenuButton.stories.js as follows, and then testing both
mouse and keyboard:

export const Default = (args) => {
  return (
    <>
      <MenuButton {...args} onClick={action('onClick')} label="Actions">
        <MenuItem
          label="First action with a long label description"
          onClick={() => { document.getElementById("my-input").focus(); }}
        />
        <MenuItem label="Second action" onClick={action('onClick')} />
        <MenuItem label="Third action" onClick={action('onClick')} disabled />
      </MenuButton>
      <input id="my-input"/>
    </>
  );
};

See related #8829, which added the initial code for closing the menu on keyup
rather than keydown.

* fix(MenuButton): add keyboard test cases

* fix(MenuButton): only use react event handlers

* fix(MenuButton): tweak comments

* fix(MenuButton): enhance test to make sure onClick is called

* fix(MenuButton): fix theoretical race condition