fix(Menu): treat Tab key like Escape
Address problems when users press the Tab key when on a dropdown
menu item, and focus ends up going somewhere weird.
Especially in the case when the menu is in a Modal/Tearsheet.

Updated both MenuButton and OverflowMenu.

The Tab key should arguably move focus to the element after the MenuButton/
OverflowMenu, or for shift-tab, the element before.  But to keep things simple,
I just return focus to the OverflowMenu/MenuButton itself.  Directing focus
before or after the MenuButton would be hard/easy depending on whether the
menu was rendered using portalling or the fixed strategy.  See
https://floating-ui.com/docs/misc#clipping for details.

About tests, it would be good to add Playwright tests for this, but I didn't
because:

1. yarn run test:e2e doesn't run any of the relevant tests anyway, it only
   runs *-test.js files whereas the relevant files have names like
   OverflowMenu-test.e2e.js.

2. As Taylor wrote in #19120, the Menu keyboard test Menu-test.avt.e2e.js
   is currently disabled.

Also:

* Deprecated focusTrap.  I think it's no longer (or never was) necessary.

* Fix race condition preventing focus when opening submenus (with the
  right arrow key).  For some reason, when the user used the right-arrow
  to open a submenu, it called focus() before the menu was actually
  displayed, thus failing.  See the menu example in Storybook (and test
  it with keyboard rather than mouse).

Fixes #17983.