fix(AILabel): fix types
Fix some of the types for Toggletip and AILabel.

In particular make Typescript know about the autoAlignBoundary
parameter.

Note though that AILabel's revertActive property is confusing as we are
either extending IconButton or Toggletip, and they are quite different.

Also, I had to jump through hoops to get the PropTypes working, in order
to avoid Typescript errors.  It seems like the *.d.ts files are generated
from the propTypes declarations rather than the Typescript declarations
(like they should be).

Refs #17232.