fix(slider): Disallow invalid & warn when readonly or disabled (#20784)
* docs: add contributor

* fix(slider): fix invalid warning when disabled (#20731)

fix slider to not show warning message when component is disabled or readOnly.

* test(slider): add tests for slider warning fixes

add test cases to not show warnings when disabled or readOnly.

* fix(slider): change variable name

change variable name from "isCurrentlyEditable" to "isInteractive".

* test(slider): add missing validations

* fix(slider): fix invalid warning when disabled (web-components)

fix slider to not show warning message when component is disabled or readOnly.

* test(slider): add tests for slider warning fixes

add test cases to not show warnings when disabled or readOnly (web-components).

* test(slider): improve tests for slider warning fixes

improve tests for two handle slider to check for data-invalid prop.

* refactor(slider): use normalized props hook

refactor slider to use the useNormalizedInputProps hook.

* test(slider): add test cases for aria-invalid

---------

Co-authored-by: Riddhi Bansal <41935566+riddhybansal@users.noreply.github.com>