fix: typescript declarations for Slider and SliderSkeleton
Those components do not have a property named "propTypes",
but the declarations in Slider.d.ts and Slider.Skeleton.d.ts
claimed they did.

I think the problem happens because there's some magic that tries
to generate Typescript declarations based on the PropTypes, and
that is fragile.  I followed the example of Row.tsx to avoid /
workaround that issue.

Fixes #20832.