Fix(time-picker): prevent invalid/warn states when disabled/readonly (#20818)
* fix(time-picker): prevent invalid/warn states when disabled/readonly

Fixed Timepicker components in both React and Web Components to not show
invalid/warn states when in disabled or readonly state, as users cannot
interact with these components in those states.

* fix(time-picker): prevent invalid/warn states when disabled/readonly

Fixed Timepicker components in both React and Web Components to not show
invalid/warn states when in disabled or readonly state, as users cannot
interact with these components in those states.

---------

Co-authored-by: Alison Joseph <alison.joseph@us.ibm.com>